refactor: estrutura livros
