--- src/main/java/br/com/ghdpreto/api_livraria/adapters/repositories/livro/SpringDataLivroRepository.java
@@ -1,11 +1,0 @@
-package br.com.ghdpreto.api_livraria.adapters.repositories.livro;
-
-import java.util.UUID;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import br.com.ghdpreto.api_livraria.frameworks.database.entities.LivroEntity;
-
-public interface SpringDataLivroRepository extends JpaRepository<LivroEntity, UUID> {
-
-}
