--- src/main/java/br/com/ghdpreto/api_livraria/application/livro/dtos/PaginaDTO.java
@@ -1,14 +1,0 @@
-package br.com.ghdpreto.api_livraria.application.livro.dtos;
-
-import java.util.List;
-
-public record PaginaDTO<T>(
-                List<T> conteudo,
-                Integer paginaAtual,
-                Integer tamanhoPagina,
-                Long totalElementos,
-                Integer totalPaginas
-
-) {
-
-}
