--- src/main/java/br/com/ghdpreto/api_livraria/application/livro/exceptions/NaoEncontradoException.java
@@ -1,7 +1,0 @@
-package br.com.ghdpreto.api_livraria.application.livro.exceptions;
-
-public class NaoEncontradoException extends RuntimeException {
-    public NaoEncontradoException(String mensagem) {
-        super(mensagem);
-    }
-}
