+++ src/main/java/br/com/ghdpreto/api_livraria/common/dto/PaginacaoDTO.java
@@ -1,0 +1,16 @@
+package br.com.ghdpreto.api_livraria.common.dto;
+
+public record PaginacaoDTO(
+        Integer pagina,
+        Integer tamanho) {
+
+    public PaginacaoDTO {
+        if (pagina == null) {
+            pagina = 0;
+        }
+        if (tamanho == null) {
+            tamanho = 10;
+        }
+    }
+
+}
