--- src/main/java/br/com/ghdpreto/api_livraria/domain/entities/Cliente.java
@@ -1,40 +1,0 @@
-package br.com.ghdpreto.api_livraria.domain.entities;
-
-import java.util.List;
-import java.util.UUID;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "cliente")
-@AllArgsConstructor(access = AccessLevel.PACKAGE)
-@NoArgsConstructor(access = AccessLevel.PACKAGE)
-@Builder(toBuilder = true)
-public class Cliente {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.UUID)
-    private UUID id;
-
-    @Column(name = "ds_nome", nullable = false)
-    private String nome;
-
-    @Column(name = "ds_email", nullable = false)
-    private String email;
-
-    @Column(name = "ds_endereco", nullable = false)
-    private String endereco;
-
-    @OneToMany(mappedBy = "cliente")
-    private List<Pedido> pedidos;
-}
