--- src/main/java/br/com/ghdpreto/api_livraria/domain/entities/Pedido.java
@@ -1,54 +1,0 @@
-package br.com.ghdpreto.api_livraria.domain.entities;
-
-import java.util.Date;
-
-import java.util.UUID;
-
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.JoinTable;
-import jakarta.persistence.ManyToMany;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Temporal;
-import jakarta.persistence.TemporalType;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PACKAGE)
-@AllArgsConstructor(access = AccessLevel.PACKAGE)
-@Builder
-public class Pedido {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.UUID)
-    private UUID id;
-
-    @ManyToOne
-    @JoinColumn(name = "id_cliente", nullable = false)
-    private Cliente cliente;
-
-    /*
-     * @ManyToMany(cascade = CascadeType.ALL, fetch = FetchType.LAZY)
-     * 
-     * @JoinTable(name = "pedido_livro", joinColumns = @JoinColumn(name =
-     * "id_pedido"), inverseJoinColumns = @JoinColumn(name = "id_livro"))
-     * private Set<Livro> livros;
-     */
-
-    @Temporal(TemporalType.TIMESTAMP)
-    @Column(name = "dt_pedido", nullable = false)
-    private Date dataPedido;
-
-    @Column(name = "vl_total", nullable = false)
-    private Double valorTotal;
-
-}
