+++ src/main/java/br/com/ghdpreto/api_livraria/modules/livro/domain/model/Livro.java
@@ -1,0 +1,60 @@
+package br.com.ghdpreto.api_livraria.modules.livro.domain.model;
+
+import java.time.LocalDateTime;
+import java.util.UUID;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.PrePersist;
+import jakarta.persistence.PreUpdate;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "livros")
+@AllArgsConstructor(access = AccessLevel.MODULE)
+@NoArgsConstructor(access = AccessLevel.MODULE)
+@Builder(toBuilder = true)
+@Getter
+public class Livro {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.UUID)
+    private UUID id;
+
+    @Column(name = "ds_titulo", nullable = false)
+    private String titulo;
+
+    @Column(name = "ds_author", nullable = false)
+    private String autor;
+
+    @Column(name = "vl_preco", nullable = false)
+    private Double preco;
+
+    @Column(name = "qt_estoque")
+    private Integer quantidadeEstoque;
+
+    @Column(name = "dt_criacao")
+    private LocalDateTime dataCriacao;
+
+    @Column(name = "dt_atualizacao")
+    private LocalDateTime dataAtualizacao;
+
+    @PrePersist
+    protected void onCreate() {
+        dataCriacao = LocalDateTime.now();
+        dataAtualizacao = LocalDateTime.now();
+    }
+
+    @PreUpdate
+    protected void onUpdate() {
+        dataAtualizacao = LocalDateTime.now();
+    }
+}
